requirements.txt
google-generativeai
moviepy
edge-tts
instagrapi
pillow
requests
screeninfo
undetected-chromedriver
selenium
